Added main menu
