+++ Assembly-CSharp.csproj
@@ -52,19 +52,20 @@
     <ImplicitlyExpandDesignTimeFacades>false</ImplicitlyExpandDesignTimeFacades>
   </PropertyGroup>
   <ItemGroup>
+     <Compile Include="Assets\Scripts\Parts\Thruster.cs" />
      <Compile Include="Assets\Scripts\Parts\Weapon.cs" />
      <Compile Include="Assets\Scripts\Parts\SpaceshipPart.cs" />
      <Compile Include="Assets\Scripts\AsteroidBehaviour.cs" />
-     <Compile Include="Assets\Scripts\GameManager.cs" />
-     <Compile Include="Assets\Scripts\SceneChanger.cs" />
+     <Compile Include="Assets\Scripts\Control\GameManager.cs" />
+     <Compile Include="Assets\Scripts\Control\SceneChanger.cs" />
      <Compile Include="Assets\Scripts\DragandDrop.cs" />
      <Compile Include="Assets\Scripts\Spaceship.cs" />
      <Compile Include="Assets\Scripts\Projectile.cs" />
-     <Compile Include="Assets\Scripts\Startup.cs" />
-     <Compile Include="Assets\Scripts\AsteroidManager.cs" />
+     <Compile Include="Assets\Scripts\Control\Startup.cs" />
+     <Compile Include="Assets\Scripts\Control\AsteroidManager.cs" />
      <Compile Include="Assets\Scripts\CreatePart.cs" />
      <Compile Include="Assets\Scripts\Background.cs" />
-     <Compile Include="Assets\Scripts\Parts\Thruster.cs" />
+     <Compile Include="Assets\Scripts\Control\MainMenu.cs" />
      <None Include="Assets\2D Space Kit\Backgrounds\Background Tip.txt" />
      <None Include="Assets\2D Space Kit\Readme.txt" />
      <None Include="Assets\2D Space Kit\Background Elements\Planets\Planet Masks\Note.txt" />
