--- Assets/Prefabs/Body.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 303a8b9a0c9e7ae4886d479a5fe1e554
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
