--- Assets/Prefabs/Spaceship.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9c68055da6d7f884f91e670d09df59c3
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
