--- Assets/Prefabs/Thruster.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5fdee65fb15cf48468829f63b26f3635
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
