--- Assets/Prefabs/Weapon.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5f9bb0543f8983840933b36c0afa2688
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
