+++ Assets/Scripts/AsteroidBehaviour.cs
@@ -1,3 +1,4 @@
+using Control;
 using Unity.Collections;
 using UnityEngine;
 using Random = UnityEngine.Random;
