+++ Assets/Scripts/Parts/Weapon.cs
@@ -1,3 +1,4 @@
+using Control;
 using UnityEngine;
 
 namespace Parts
