+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,9 +5,6 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 0
-    path: 
-    guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/BuildingScene.unity
     guid: 1628852b1ff572d40ade77b4353110e2
