+++ UserSettings/EditorUserSettings.asset
@@ -9,6 +9,9 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c791c0f1e162d1f2b233e2867083debf42d
       flags: 0
     RecentlyUsedScenePath-1:
+      value: 22424703114646680e0b0227036c6c111b07142f1f2b233e2867083debf42d
+      flags: 0
+    RecentlyUsedScenePath-2:
       value: 22424703114646680e0b0227036c7d051f1b1c23222f15332827187df7ee3d2cfb
       flags: 0
     vcSharedLogLevel:
